feat(snapshot): move file load path to snapshot storage
